replace u32 with more generic uint_32 type, clarify some type cast conversions

+++ Bme280BoschWrapper.cpp
@@ -68,12 +68,12 @@ int32_t Bme280BoschWrapper::getTemperature()
   return comp_data.temperature;
 }
 
-u32 Bme280BoschWrapper::getHumidity()
+uint32_t Bme280BoschWrapper::getHumidity()
 {
   return comp_data.humidity;
 }
 
-u32 Bme280BoschWrapper::getPressure()
+uint32_t Bme280BoschWrapper::getPressure()
 {
   return comp_data.pressure;
 }
@@ -119,17 +119,17 @@ int8_t Bme280BoschWrapper::I2CRead(uint8_t dev_addr, uint8_t reg_addr, uint8_t *
   Wire.write(reg_addr);
   Wire.endTransmission();
   
-  Wire.requestFrom(dev_addr, cnt);
+  Wire.requestFrom(dev_addr, (uint8_t) cnt);
   
   uint8_t available = Wire.available();
-  if(available != cnt)
+  if(available != (uint8_t) cnt)
   {
     ret = BME280_E_COMM_FAIL;
   }
   
   for(uint8_t i = 0; i < available; i++)
   {
-    if(i < cnt) 
+    if(i < (uint8_t) cnt) 
     {
       *(reg_data + i) = Wire.read();
 //      Serial.print(*(reg_data + i), HEX);
@@ -213,7 +213,7 @@ int8_t Bme280BoschWrapper::SPIWrite(uint8_t dev_addr, uint8_t reg_addr, uint8_t 
   return ret;
 }
 
-void Bme280BoschWrapper::delaymsec(u32 msec)
+void Bme280BoschWrapper::delaymsec(uint32_t msec)
 {
   delay(msec);
 }
